bundle update and fix errors caused by Devise & Select2 changes
Devise added another mailer method, which upset the mailer_previews
test. Changed the mailer_previews test so it will only care about
previews for Devise mails that are actually implemented in the app.

Select2 changed something subtle about when the result options in the
dropdown render. It seems to fix itself if you wait for the option's
text to actually appear before trying to click on it.